chore: fixed dependency issue 2
